src/requirements.txt
additional-urwid-widgets==0.4.1
git+https://github.com/PyCQA/astroid@master
coloredlogs==14.0
certifi==2020.4.5.1
chardet==3.0.4
cloudant==2.12.0
dataclasses==0.6
executor==22.0
fasteners==0.15
graphql-core==3.1
graphqlclient==0.2.4
humanfriendly>=8.2
idna==2.9
isort==4.3.21
jsonpickle==1.3
lazy-object-proxy>=1.5.1
mccabe==0.6.1
monotonic==1.5
#pkg-resources==0.0.0
git+https://github.com/AadamZ5/python-proc.git@dev
git+https://github.com/AadamZ5/py-controller-model@dev
property-manager==3.0
py-ts-interfaces==0.2.0
git+https://github.com/freenas/py-SMART@master
pyudev==0.22.0
sgqlc==10.1
six==1.14.0
urllib3==1.25.8
urwid==2.1.0
verboselogs==1.7
websockets==8.1
wrapt==1.11.2
injectable==3.4.4
typing-inspect==0.6.0
cached-property==1.5.1
pycollect==0.2.3
parameters-validation==1.2.0
typing-extensions>=3.7.4
mypy-extensions>=0.3.0
aniso8601==7.0.0
astroid==2.5.0
click==7.1.2
controllermodel==0.1
graphql-relay==2.0.1
h11==0.12.0
hupper==1.10.2
proc==1.0
promise==2.3
Pygments==2.7.3
pySMART==0.3
python-dateutil==2.8.1
requests==2.25.1
Rx==1.6.1
starlette==0.14.1
strawberry-graphql==0.44.6
uvicorn==0.13.3